+++ src/lib/index.ts
@@ -2,7 +2,7 @@ export { default as AnchorFm } from './components/anchor-fm.svelte'
 export { default as Buzzsprout } from './components/buzzsprout.svelte'
 export { default as Deezer } from './components/deezer.svelte'
 export { default as GeneralObserver } from './components/general-observer.svelte'
-export { default as GenericEmbed } from './components/generic.svelte'
+export { default as GenericEmbed } from './components/generic-embed.svelte'
 export { default as SimpleCast } from './components/simple-cast.svelte'
 export { default as SoundCloud } from './components/sound-cloud.svelte'
 export { default as Spotify } from './components/spotify.svelte'
